backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
supabase>=2.11.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18
httpx>=0.28.0
Pillow>=11.0.0
google-genai>=1.0.0
requests>=2.32.0
eval-type-backport>=0.2.0
